requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
aiohttp==3.9.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1 